--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -14,16 +14,10 @@ type DomainStat map[string]int
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 	result := make(DomainStat)
 	bufReader := bufio.NewReader(r)
-	var email string
 	var flag bool
-	var matched bool
-	var err error
-	var num int
-	var domainstr string
-	var line string
 	searchDomain := "." + domain
 	for {
-		line, err = bufReader.ReadString('\n')
+		line, err := bufReader.ReadString('\n')
 		if err != nil {
 			if errors.Is(err, io.EOF) {
 				flag = true
@@ -31,13 +25,13 @@ func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 				return nil, err
 			}
 		}
-		email = gjson.Get(line, "Email").String()
+		email := gjson.Get(line, "Email").String()
 
-		matched = strings.Contains(email, searchDomain)
+		matched := strings.Contains(email, searchDomain)
 
 		if matched {
-			domainstr = strings.ToLower(strings.SplitN(email, "@", 2)[1])
-			num = result[domainstr]
+			domainstr := strings.ToLower(strings.SplitN(email, "@", 2)[1])
+			num := result[domainstr]
 			num++
 			result[domainstr] = num
 		}