--- a/hw10_program_optimization/stats_test.go
+++ b/hw10_program_optimization/stats_test.go
@@ -5,10 +5,18 @@ package hw10programoptimization
 
 import (
 	"bytes"
-	"fmt"
+	//"fmt"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/require"
+)
+
+const (
+	mb          uint64 = 1 << 20
+	memoryLimit uint64 = 30 * mb
+
+	timeLimit = 300 * time.Millisecond
 )
 
 func TestGetDomainStat(t *testing.T) {
@@ -38,8 +46,25 @@ func TestGetDomainStat(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, DomainStat{}, result)
 	})
+	t.Run("AdditionalBenchmark", func(t *testing.T) {
+		bench := func(b *testing.B) {
+			for i := 1; i < b.N; i++ {
+				res, _ := GetDomainStat(bytes.NewBufferString(data), "com")
+				b.Helper()
+				b.StopTimer()
+				//fmt.Println(result)
+				require.NotEqual(t, res, DomainStat{})
+				b.StartTimer()
+			}
+		}
+		result := testing.Benchmark(bench)
+		mem := result.MemBytes
+		t.Logf("time used: %s / %s", result.T, timeLimit)
+		t.Logf("memory used: %dMb / %dMb", mem/mb, memoryLimit/mb)
+	})
 }
 
+/*
 func AdditionalBenchmark(b *testing.B) {
 
 	data := `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
@@ -54,3 +79,26 @@ func AdditionalBenchmark(b *testing.B) {
 		}
 	})
 }
+
+func AdditionalBenchmark2(t *testing.T) {
+	data := `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
+{"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"mLynch@broWsecat.com","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}
+{"Id":3,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
+{"Id":4,"Name":"Gregory Reid","Username":"tButler","Email":"5Moore@Teklist.net","Phone":"520-04-16","Password":"r639qLNu","Address":"Sunfield Park 20"}
+{"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@Linktype.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}`
+
+	bench := func(b *testing.B) {
+		for i := 1; i < b.N; i++ {
+			_, _ = GetDomainStat(bytes.NewBufferString(data), "com")
+			b.Helper()
+			b.StopTimer()
+			//fmt.Println(result)
+			b.StartTimer()
+		}
+	}
+	result := testing.Benchmark(bench)
+	mem := result.MemBytes
+	t.Logf("time used: %s / %s", result.T, timeLimit)
+	t.Logf("memory used: %dMb / %dMb", mem/mb, memoryLimit/mb)
+}
+*/