## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [ ] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [ ] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов

Optimization benchmark before optimization:
==============================================================================
=== RUN   TestGetDomainStat_Time_And_Memory
    stats_optimization_test.go:45: time used: 472.7265ms / 300ms
    stats_optimization_test.go:46: memory used: 308Mb / 30Mb
    assertion_compare.go:332: 
        	Error Trace:	stats_optimization_test.go:48
        	Error:      	"472726500" is not less than "300000000"
        	Test:       	TestGetDomainStat_Time_And_Memory
        	Messages:   	[the program is too slow]
--- FAIL: TestGetDomainStat_Time_And_Memory (11.43s)
FAIL
exit status 1
FAIL	github.com/skolzkyi/hwOTUS_YIA/hw10_program_optimization	15.801s
===============================================================================
Additional benchmark before optimization:
===============================================================================
allocated bytes per operation: 134685660 B/op
allocations per operations: 1683089 allocs/op
PASS
ok  	github.com/skolzkyi/hwOTUS_YIA/hw10_program_optimization	16.364s
===============================================================================
Optimization benchmark after optimization:
===============================================================================
=== RUN   TestGetDomainStat_Time_And_Memory
    stats_optimization_test.go:46: time used: 111.248ms / 300ms
    stats_optimization_test.go:47: memory used: 17Mb / 30Mb
--- PASS: TestGetDomainStat_Time_And_Memory (1.12s)
PASS
ok  	github.com/skolzkyi/hwOTUS_YIA/hw10_program_optimization	3.509s
===============================================================================
Additional benchmark after optimization:
===============================================================================
allocated bytes per operation: 6288 B/op
allocations per operations: 16 allocs/op
PASS
ok  	github.com/skolzkyi/hwOTUS_YIA/hw10_program_optimization	22.104s
===============================================================================