@antonpriyma 2023-03-08
Я бы еще все таки не объявлял переменные до цикла (num, err, domainStr и тд). На аллокации это почти не повлияет, а в продакшн коде так непринято делать, снижается читаемость. Но это минорно и можно не считать замечанием)